requirements.txt
flask
yfinance
plotly
pandas
requests
python-dotenv